Move geoencoder to /political/geoencoder
