--- app/routes/thematic/__init__.py
@@ -1,2 +1,0 @@
-"""Here live a number of endpoints which provide higher-level services
-mostly intended to make life easier for consumers of the Data API."""
